move and rename files
This is a small organizational change that gives each component its own directory inside components/ with index.tsx and styles.css as standard.
